requirements.txt
langgraph==0.0.30
langchain==0.1.4
langchain-core==0.1.16
langchain-openai==0.0.5
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-weather==1.0.3
alphavantage==0.0.6
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0
